make the project more structured
